requirements.txt
--find-links https://download.pytorch.org/whl/cu113/torch_stable.html
torch>=1.10.1+cu113
torchvision>=0.11.2+cu113
torchmetrics>=0.6.2
albumentations>=1.1.0
madgrad>=1.1
matplotlib>=3.5.1
matplotlib-inline>=0.1.3
mlflow>=1.22.0
numpy>=1.21.4
pandas>=1.3.4
Pillow>=8.4.0
pytorch-lightning>=1.5.8
scikit-image>=0.19.1
scikit-learn>=1.0.1
scipy>=1.7.3
timm>=0.4.12
pytorchvideo>=0.1.5